Move /config_xml view to the views/canvas namespace
