+++ lms/routes.py
@@ -11,7 +11,7 @@ def includeme(config):
     config.add_route("logout", "/logout")
     config.add_route("reports", "/reports")
 
-    config.add_route("config_xml", "/config_xml")
+    config.add_route("canvas.v11.config", "/config_xml")
     config.add_route(
         "configure_assignment",
         "/assignment",
