+++ lms/views/canvas/config.py
@@ -2,12 +2,12 @@ from pyramid.view import view_config
 
 
 @view_config(
-    route_name="config_xml",
+    route_name="canvas.v11.config",
     renderer="lms:templates/config.xml.jinja2",
     request_method="GET",
 )
 def config_xml(request):
-    """Render the XML configuration as XML."""
+    """Render the XML configuration."""
     request.response.content_type = "text/xml"
 
     return {
