+++ tests/unit/lms/views/canvas/config_test.py
@@ -1,4 +1,4 @@
-from lms.views.config import config_xml
+from lms.views.canvas.config import config_xml
 
 
 class TestConfigXml:
